Move Kobold Classic settings into related module
