+++ public/scripts/RossAscends-mods.js
@@ -4,7 +4,6 @@ import {
     characters,
     online_status,
     main_api,
-    api_server,
     is_send_press,
     max_context,
     saveSettingsDebounced,
@@ -42,6 +41,7 @@ import { debounce_timeout } from './constants.js';
 import { Popup } from './popup.js';
 import { accountStorage } from './util/AccountStorage.js';
 import { getCurrentUserHandle } from './user.js';
+import { kai_settings } from './kai-settings.js';
 
 var RPanelPin = document.getElementById('rm_button_panel_pin');
 var LPanelPin = document.getElementById('lm_button_panel_pin');
@@ -371,7 +371,7 @@ function RA_autoconnect(PrevApi) {
     if (online_status === 'no_connection' && power_user.auto_connect) {
         switch (main_api) {
             case 'kobold':
-                if (api_server && isValidUrl(api_server)) {
+                if (kai_settings.api_server && isValidUrl(kai_settings.api_server)) {
                     $('#api_button').trigger('click');
                 }
                 break;
