+++ public/scripts/preset-manager.js
@@ -598,6 +598,7 @@ class PresetManager {
         }
 
         const filteredKeys = [
+            'api_server',
             'preset',
             'streaming',
             'truncation_length',
@@ -607,6 +608,7 @@ class PresetManager {
             'can_use_tokenization',
             'can_use_streaming',
             'preset_settings_novel',
+            'preset_settings',
             'streaming_novel',
             'nai_preamble',
             'model_novel',
